fixing conflicting files
