+++ wordle/src/HomePage/HomePage.scss
@@ -1,0 +1,4 @@
+.home-style {
+    position: relative;
+  }
+
